Refactor <ScanningJobPanel/> for removing jobs
- Move the panel body in a separate component, <ScannigJobPanelBody/>
- Add state to the panel to decide what is shown inside (normal body or
    remove dialogue).
- Move the status label logic in it's own conponent as well
