+++ scanomatic/ui_server_data/js/src/containers/ScanningRootContainer.jsx
@@ -3,7 +3,7 @@ import React from 'react';
 import { getScanningJobs, startScanningJob, deleteScanningJob } from '../api';
 import { getScannersWithOwned } from '../helpers';
 import ScanningRoot from '../components/ScanningRoot';
-import { duration2milliseconds } from '../components/ScanningJobPanel';
+import { duration2milliseconds } from '../components/ScanningJobPanelBody';
 
 
 export default class ScanningRootContainer extends React.Component {
